Refactor: move combinations function to e2.py and update references in e3.py; remove w2.py and w3.py
